\name{indicatorQogBasic}
\alias{indicatorQogBasic}
\title{List the currently available
indicators from Quality of Government Institute Basic data}
\usage{
  indicatorQogBasic(urli = "http://www.qogdata.pol.gu.se/data/qog_bas_ts_30aug13.csv")
}
\arguments{
  \item{urli}{is the url address to the data}
}
\value{
  string of names of available indicators
}
\description{
  List the currently available indicators from Quality of
  Government Institute Basic data
  \url{http://www.qog.pol.gu.se/data/datadownloads/qogbasicdata/}
}
\examples{
# available.indicators <- indicatorQogBasic()
}
\author{
  Markus Kainu <markuskainu(at)gmail.com>
}

