\name{indicatorQogStandard}
\alias{indicatorQogStandard}
\title{List the currently available
indicators from Quality of Government Institute Standard data}
\usage{
  indicatorQogStandard(urli = "http://www.qogdata.pol.gu.se/data/qog_std_ts_15may13.csv")
}
\arguments{
  \item{urli}{is the url address to the data}
}
\value{
  string of names of available indicators
}
\description{
  List the currently available indicators from Quality of
  Government Institute Standard data
  \url{http://www.qog.pol.gu.se/data/datadownloads/qogstandarddata/}
}
\examples{
# available.indicators <- indicatorQogStandard()
}
\author{
  Markus Kainu <markuskainu(at)gmail.com>
}

