% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_basic_cs_2020}
\alias{meta_basic_cs_2020}
\title{Metadata for 2020 Quality of Government institute Basic Data - cross-sectional}
\format{A data frame with 590 rows and five columns 
\describe{
   \item{code}{Indicator code in data}
   \item{name}{Name of the indicator}
   \item{value}{Indicator value for factor indicator}
   \item{label}{Value label for factor indicators}
   \item{class}{Class of the indicator. Numeric, factor or character}
}}
\usage{
meta_basic_cs_2020
}
\description{
Generated automatically from spss-version of Quality of Government Basic data using \code{tidymetadata::create_metadata()} function. 
Formatted in tidy-format with row per each value of an indicator. Numeric indicators have a single row per indicator.
}
\keyword{datasets}
