% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_oecd_cs_2016}
\alias{meta_oecd_cs_2016}
\title{Metadata for 2016 Quality of Government institute OECD Data - cross-sectional}
\format{
A data frame with 1608 rows and five columns 
\describe{
   \item{code}{Indicator code in data}
   \item{name}{Name of the indicator}
   \item{value}{Indicator value for factor indicator}
   \item{label}{Value label for factor indicators}
   \item{class}{Class of the indicator. Numeric, factor or character}
}
}
\usage{
meta_oecd_cs_2016
}
\description{
Generated automatically from spss-version of Quality of Government OECD data using \code{tidymetadata::create_metadata()} function. 
Formatted in tidy-format with row per each value of an indicator. Numeric indicators have a single row per indicator.
}
\seealso{
Other datasets: 
\code{\link{meta_basic_cs_2016}},
\code{\link{meta_basic_cs_2017}},
\code{\link{meta_basic_cs_2018}},
\code{\link{meta_basic_cs_2019}},
\code{\link{meta_basic_cs_2020}},
\code{\link{meta_basic_cs_2021}},
\code{\link{meta_basic_cs_2022}},
\code{\link{meta_basic_cs_2023}},
\code{\link{meta_basic_ts_2016}},
\code{\link{meta_basic_ts_2017}},
\code{\link{meta_basic_ts_2018}},
\code{\link{meta_basic_ts_2019}},
\code{\link{meta_basic_ts_2020}},
\code{\link{meta_basic_ts_2021}},
\code{\link{meta_basic_ts_2023}},
\code{\link{meta_oecd_cs_2017}},
\code{\link{meta_oecd_cs_2018}},
\code{\link{meta_oecd_cs_2019}},
\code{\link{meta_oecd_cs_2020}},
\code{\link{meta_oecd_cs_2021}},
\code{\link{meta_oecd_cs_2022}},
\code{\link{meta_oecd_cs_2023}},
\code{\link{meta_oecd_ts_2016}},
\code{\link{meta_oecd_ts_2017}},
\code{\link{meta_oecd_ts_2018}},
\code{\link{meta_oecd_ts_2019}},
\code{\link{meta_oecd_ts_2020}},
\code{\link{meta_oecd_ts_2021}},
\code{\link{meta_oecd_ts_2022}},
\code{\link{meta_oecd_ts_2023}},
\code{\link{meta_std_cs_2016}},
\code{\link{meta_std_cs_2017}},
\code{\link{meta_std_cs_2018}},
\code{\link{meta_std_cs_2019}},
\code{\link{meta_std_cs_2020}},
\code{\link{meta_std_cs_2021}},
\code{\link{meta_std_cs_2022}},
\code{\link{meta_std_cs_2023}},
\code{\link{meta_std_ts_2016}},
\code{\link{meta_std_ts_2017}},
\code{\link{meta_std_ts_2018}},
\code{\link{meta_std_ts_2019}},
\code{\link{meta_std_ts_2020}},
\code{\link{meta_std_ts_2021}},
\code{\link{meta_std_ts_2022}},
\code{\link{meta_std_ts_2023}}
}
\concept{datasets}
\keyword{datasets}
