% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_oecd_ts_2018}
\alias{meta_oecd_ts_2018}
\title{Metadata for 2018 Quality of Government institute Standard Data - time-series}
\format{A data frame with 1407 rows and five columns 
\describe{
   \item{code}{Indicator code in data}
   \item{name}{Name of the indicator}
   \item{value}{Indicator value for factor indicator}
   \item{label}{Value label for factor indicators}
   \item{class}{Class of the indicator. Numeric, factor or character}
}}
\usage{
meta_oecd_ts_2018
}
\description{
Generated automatically from spss-version of Quality of Government OECD data using \code{tidymetadata::create_metadata()} function. 
Formatted in tidy-format with row per each value of an indicator. Numeric indicators have a single row per indicator.
}
\keyword{datasets}
