% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run an on-board Shiny app to browse the metadata and data availability of Quality of government data}
\usage{
run_app()
}
\description{
Run an on-board Shiny app to browse the metadata and data availability of Quality of government data
}
\examples{
# run_app()
}
\author{
Markus Kainu <markus.kainu(at)kapsi.fi>
}
