% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qog.R
\name{read_qog}
\alias{read_qog}
\title{Retrieve and Read Quality of Government Institute data in R}
\usage{
read_qog(which.data = "basic", data.type = "time-series",
  data.dir = "~/tmp")
}
\arguments{
\item{which.data}{A string. Specify the name of the QoG data set to retrieve. Currently available \code{"basic"}, \code{"standard"}, \code{"oecd"} or \code{"social_policy"}.}

\item{data.type}{A string. Specify whether you want cross-sectional or time-series QoG data set to retrieve. Currently available \code{"cross-sectional"} or \code{"time-series"}.}

\item{data.dir}{A string. Specify the path where to save the downloaded data file.}
}
\value{
data.frame
}
\description{
Download data from Quality of Government Institute Basic data at
\url{http://www.qog.pol.gu.se/data/}
}
\examples{
# dat <- read_qog(which.data = "basic", data.dir="data")
}
\author{
Markus Kainu <markuskainu(at)gmail.com>
}
