% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qog.R
\name{read_qog}
\alias{read_qog}
\title{Retrieve and Read Quality of Government Institute data in R}
\usage{
read_qog(
  which_data = "basic",
  data_type = "time-series",
  year = 2022,
  data_dir = NULL,
  file_format = "csv",
  download_only = FALSE,
  cache = TRUE,
  update_cache = FALSE
)
}
\arguments{
\item{which_data}{A string. Specify the name of the QoG data set to retrieve. Currently available \code{"basic"}, \code{"standard"}, \code{"oecd"} or \code{"social_policy"}.}

\item{data_type}{A string. Specify whether you want cross-sectional or time-series QoG data set to retrieve. Currently available \code{"cross-sectional"} or \code{"time-series"}.}

\item{year}{numeric. Specify the year of the dataset. Year refers to the publication year of the dataset not the year of a particular data point. Default \code{2019}}

\item{data_dir}{a path to a cache directory. The directory have to exist.
The \code{NULL} (default) uses and creates
'rqog' directory in the temporary directory from
\code{\link{tempdir}}.}

\item{file_format}{A string. Specify the file format you want to download and import. Currently available \code{"csv"},\code{"dta"}, \code{"sav"} or \code{"xlsx"}.}

\item{download_only}{a logical whether to only download and save the dataset and skip importing in R.  Default is \code{FALSE}.}

\item{cache}{a logical whether to do caching. Default is \code{TRUE}.}

\item{update_cache}{a locigal whether to update cache. Default is \code{FALSE}.}
}
\value{
data.frame
}
\description{
Download data from Quality of Government Institute Basic data at
\url{http://www.qog.pol.gu.se/data/}
}
\examples{
# dat <- read_qog(which_data = "basic", year = 2019, data_dir="data")
}
\author{
Markus Kainu <markuskainu(at)gmail.com>
}
